requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Redis for queuing
redis==5.0.1
aioredis==2.0.1

# External service providers
sendgrid==6.10.0
firebase-admin==6.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.0.3

# HTTP client for testing
httpx==0.25.2

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0